backend/pdf-graphrag-service/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart
PyMuPDF==1.23.7
openai==1.99.6
neo4j==5.20.0
python-dotenv
requests
